conf: 패키지 이동
