+++ src/main/java/com/example/fieldpasserbe/seok/entity/TransactionStatus.java
@@ -1,4 +1,4 @@
-package com.example.fieldpasserbe.entity;
+package com.example.fieldpasserbe.seok.entity;
 
 public enum TransactionStatus {
     판매중,예약중,판매완료;
